+++ vcf/src/vcf/GraphicsKit/Win32GraphicsToolkit.h
@@ -25,7 +25,7 @@ public:
 
 	virtual ~Win32GraphicsToolkit();
 
-	virtual ContextPeer* internal_createContextPeer( const unsigned long& width, const unsigned long& height );
+	virtual ContextPeer* internal_createContextPeer( const uint32& width, const uint32& height );
 
 	virtual ContextPeer* internal_createContextPeer( OSHandleID contextID );
 
@@ -33,7 +33,7 @@ public:
 
 	virtual FontPeer* internal_createFontPeer( const String& fontName, const double& pointSize );
 
-	virtual Image* internal_createImage( const unsigned long& width, const unsigned long& height, const Image::ImageType& imageType );
+	virtual Image* internal_createImage( const uint32& width, const uint32& height, const Image::ImageType& imageType );
 
 	virtual Image* internal_createImage( GraphicsContext* context, Rect* rect, const Image::ImageType& imageType );
